proxy: Move PGLB-related modules into pglb root module.
